requirements.txt
arrow==1.2.3
asgiref==3.7.2
Babel==2.12.1
binaryornot==0.4.4
certifi==2023.7.22
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
cookiecutter==2.3.0
Django==4.2.4
django-babel==0.6.2
django-browser-reload==1.11.0
django-environ==0.10.0
django-phonenumber-field==7.1.0
django-tailwind==3.6.0
idna==3.4
Jinja2==3.1.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
phonenumbers==8.13.19
Pillow==10.0.0
Pygments==2.16.1
python-dateutil==2.8.2
python-slugify==8.0.1
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
rich==13.5.2
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
tzdata==2023.3
urllib3==2.0.4
